nordea: extend to SE/NO/DK/FI, fix ATM/depmachine categories
